+++ src/model/attention/self_attention.py
@@ -1,5 +1,5 @@
 import tensorflow as tf
-from tensorflow.keras.layers import Dense
+from tensorflow.keras.layers import Dense  # type: ignore
 
 from ...utils.positional_encoding import gen_sineembed_for_position
 
@@ -91,7 +91,9 @@ class SelfAttention(tf.keras.layers.Layer):
         v = self._proj_to_value(x_to_value)
 
         # shape = (batch_size, heads_num, seq_len, seq_len)
-        a_sc = tf.matmul(q, tf.transpose(k, perm=[0, 1, 3, 2])) / tf.sqrt(self._d_k)
+        a_sc = tf.matmul(q, tf.transpose(k, perm=[0, 1, 3, 2])) / tf.sqrt(
+            tf.cast(self._d_k, tf.float32)
+        )
         a_sc = tf.keras.activations.softmax(a_sc)
         output = tf.matmul(a_sc, v)
